requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
WTForms==3.0.1
gunicorn==21.2.0
Flask-Login==0.6.3
python-dotenv==1.0.0
Pillow==10.1.0
